requirements.txt
fastapi
uvicorn
MetaTrader5
python-dotenv
celery
redis
pydantic
pyjwt[crypto]
requests
boto3
